Migrate Character specific functions to Player_Character
