+++ Source/ACE.Server/WorldObjects/Player_Spells.cs
@@ -1,9 +1,7 @@
-using System.Collections.Generic;
 using System.Linq;
 
 using ACE.Database.Models.Shard;
 using ACE.DatLoader;
-using ACE.Entity;
 using ACE.Entity.Enum;
 using ACE.Server.Entity.Actions;
 using ACE.Server.Network.GameEvent.Events;
@@ -92,75 +90,6 @@ namespace ACE.Server.WorldObjects
                 GameEventMagicRemoveSpellId removeSpellEvent = new GameEventMagicRemoveSpellId(Session, spellId);
                 Session.Network.EnqueueSend(removeSpellEvent);
             }).EnqueueChain();
-        }
-
-
-        /// <summary>
-        /// Will return the spells in the bar, sorted by their position
-        /// </summary>
-        public List<SpellBarPositions> GetSpellsInSpellBar(int barId)
-        {
-            var spells = new List<SpellBarPositions>();
-
-            var results = Character.CharacterPropertiesSpellBar.Where(x => x.SpellBarNumber == barId);
-
-            foreach (var result in results)
-            {
-                var entity = new SpellBarPositions(result.SpellBarNumber, result.SpellBarIndex, result.SpellId);
-
-                spells.Add(entity);
-            }
-
-            spells.Sort((a, b) => a.SpellBarPositionId.CompareTo(b.SpellBarPositionId));
-
-            return spells;
-        }
-
-        /// <summary>
-        /// This method implements player spell bar management for - adding a spell to a specific spell bar (0 based) at a specific slot (0 based).
-        /// </summary>
-        public void HandleActionAddSpellFavorite(uint spellId, uint spellBarPositionId, uint spellBarId)
-        {
-            var spells = GetSpellsInSpellBar((int)spellBarId);
-
-            if (spellBarPositionId > spells.Count + 1)
-                spellBarPositionId = (uint)(spells.Count + 1);
-
-            // We must increment the position of existing spells in the bar that exist on or after this position
-            foreach (var property in Character.CharacterPropertiesSpellBar)
-            {
-                if (property.SpellBarNumber == spellBarId && property.SpellBarIndex >= spellBarPositionId)
-                    property.SpellBarIndex++;
-            }
-
-            var entity = new CharacterPropertiesSpellBar { CharacterId = Biota.Id, SpellBarNumber = spellBarId, SpellBarIndex = spellBarPositionId, SpellId = spellId };
-
-            Character.CharacterPropertiesSpellBar.Add(entity);
-            CharacterChangesDetected = true;
-        }
-
-        /// <summary>
-        /// This method implements player spell bar management for - removing a spell to a specific spell bar (0 based)
-        /// </summary>
-        public void HandleActionRemoveSpellFavorite(uint spellId, uint spellBarId)
-        {
-            var entity = Character.CharacterPropertiesSpellBar.FirstOrDefault(x => x.SpellBarNumber == spellBarId && x.SpellId == spellId);
-
-            if (entity != null)
-            {
-                // We must decrement the position of existing spells in the bar that exist after this position
-                foreach (var property in Character.CharacterPropertiesSpellBar)
-                {
-                    if (property.SpellBarNumber == spellBarId && property.SpellBarIndex > entity.SpellBarIndex)
-                    {
-                        property.SpellBarIndex--;
-                        CharacterChangesDetected = true;
-                    }
-                }
-
-                Character.CharacterPropertiesSpellBar.Remove(entity);
-                CharacterChangesDetected = true;
-            }
         }
     }
 }
